Reimplemented much of the error reporting.  Now uses a revised XMLError class, expanded and cleaned-up error diagnostic codes and parser mappings (for all three parsers: Xerces, Expat, libxml).  The XMLError class is used in the XML layer, and from this, we derive a new class called SBMLError for use in the rest of the libSBML code.  All of the basic predefined errors generated by libSBML and that were previously hardwired all over the place now have symbolic names like SBMLError::InvalidIdSyntax.  Finally, the error categories are now integer codes rather than strings.  In the same way that severity codes are defined in an enumeration, the category codes are also in an enumeration.  There are basic values defined in XMLError for the XML layer, and an expanded set of values in SBMLError for the rest of libSBML.
Not everything is revised.  The most notable case is that the validator and
the constraints defined there don't use the symbolic codes, nor indeed any
of the rest of the new machinery.  Things will work fine without that, but
at some point we should update the validator as well.

There is still quite a bit of documentation work to do.  Not all of the new
or modified classes have updated documentation comments yet.