Fixed memory leak in ModelHistory class.
ModelCreator objects that are stored in mCreators (a List object in
ModelHistory class) should be deleted by the destructor of ModelHistory 
class, because the ModelCreator objects added by addCreator function 
are cloned objects (i.e. copied by clone function) and thus the 
ModelHistory instance (not the caller) has the ownership of those objects.