Tracker issue #2182856
A CVTerm does not make sense without a metaid so the addCVTerm will not add a term if the object has no metaid.
This would be covered by the new set/add API but I think it is worth putting in a release that does not all the way. 

I added documentation about this.