This replaces the fix I did (Revision 8653) to add tougher checks to getSBMLDocument and getParentSBMLObject functions so that if the parent or document has been deleted the functions return NULL and do not crash.
My previous fix prevented crashes but resulted in inconsistent behaviour depending on the compiler used to build libSBML.

This fix seems to work for different compilers.

Each SBase object has a boolean flag mHasBeenDeleted which is false unless the destructor of the SBase object has been called ; in which case the falg is true. The getSBMLDocument and getParentSBMLObject functions check this flag (for those compilers that allow them to do so!!) and thus can return NULL if an object has been deleted even if it still exists in memory.