Imported skelton files for creating a package (.dmg file) for MacOSX. Thanks to Akira Funahashi for contributing the skelton files.
Now, running "make create-dmg" in the top directory will create a 

  libsbml-${VERSION}-${XML_PARSER}_macosx-installer.dmg

in the top directory of source tree.

Install prefix, underlying XML parser (expat, libxml2, or xerces), language
bindings and etc. are determined by options passed to a configure script.

For example, a created package will install libSBML (depends on expat) as
universal binaries (i386 and ppc) in /usr/local/ directory (default path)
with Java and Python bindings if the configure script is run as follows:

 ./configure --with-expat --enable-universal-binary --with-java --with-python