(Merge from trunk)
Imported a skeleton spec file (libsbml.spec.in) for creating libSBML RPM 
packages.

RPM packages can be created as follows.

Firstly, libsbml.spec needs to be generated by running configure script.
By default, libxml2 is used for building libSBML.
To use expat or xerces instead of libxml2, run the configure script as
follows:

 (expat)  % ./configure --with-expat
 (xerces) % ./configure --with-xerces

If you want to use Doxygen and/or SWIG which is not installed in system path
(i.e., not installed in /usr ), then add --with-doxygen=/path/to/doxygen
and/or --with-swig=/path/to/swig when running the configure script.

To create libSBML RPM packages, run rpmbuild command as follows:
(libsbml-@PACKAGE_VERSION@-src.zip needs to be located in SOURCES directory 
of your RPM working directory.)

  % rpmbuild -ba libsbml.spec

To create RPM packages of each language bindings, add the following 
--with XXXX options when running rpmbuild command:

 (Java)   --with java
 (C#)     --with csharp
 (Python) --with python
 (Ruby)   --with ruby
 (Perl)   --with perl
 (Octave) --with octave

For example, the following command will create all subpackages:

 % rpmbuild -ba libsbml.spec --with java --with csharp --with python \
             --with ruby --with perl --with octave