The flag --with-swig now acts like the other --with-foo flags: it lets the user specify where to find the desired copy of swig.  This is in contrast to what was happening before, which was that if --with-swig was given, it meant also to force reswig'ing in some cases.
If --with-swig is not given but configure detects that it
will be needed to build the bindings, it will look for swig
on the user's /Users/mhucka/system/i386-darwin98/bin /Users/mhucka/system/bin /usr/local/bin /usr/bin/X11 /usr/openwin/bin /usr/ucb /usr/bin /bin /sbin /usr/ccs/bin /usr/5bin /usr/sbin /usr/etc /etc /sw/bin /sw/sbin /Library/Frameworks/Python.framework/Versions/Current/bin /Developer/Tools /System/Library/CoreServices and test the version.

The new makefile changes basically tries to make things like
a more straightforward make dependency scheme, wherein it
always tests if libsbml_wrap.cpp is out of date and tries to
rebuild it if so.  Normally, re-swig'ing shouldn't be
necessary: as you know, in a distribution, the
libsbml_wrap.cpp will be up to date, so it shouldn't matter
to end users.  And the dependencies on source files means
that only if a source file is changed will it try to
re-swig, and that should happen only for developers who are
mucking with libsbml code (for whom it's ok to require
swig).

The tests for versions and consistency happen at both
configure time and compile time, to be (hopefully) more
robust.