Fixed a bug where in this scenario:
1   <sBaseRef>
2     <sBaseRef />
3   </sBaseRef>

the parser would think that line 2 was the closing tag of line 1 because a) it is a closing tag and b) it has the same name.  The problem is that it is also a start tag of the same name:  so, the fix is to add the check '&& !isStart()'.  Thanks to Frank for helping me track this down.

(also a couple syntactic changes)