More updates to get round-tripping working, mostly working on the instantSBML-to-coffeescript text bit.  Changed things so that if an element only has a single defined 'math' child, you get the shortcut:
   J0:
     math: "5x+y"

becomes:

   J0: "5x+y"

More of the tests now work; now working on implementing outputting the shortcut versions of the reactions, when it's appropriate to do so (i.e. when the shortcut version contains all the information in the object).

Also, fixed a bug where using '$' in the coffeescript input would sometimes stay in the attribute name, which is of course illegal in XML.

Also also, went ahead and changed some of my renamed things to their actual SBML names ('assignments' -> 'eventAssignments'; 'functions' -> 'functionDefinitions').  Might allow the old versions eventually as shortcuts, but that might actually be more confusing than simply using the SBML name.