Reorganized a lot of cut-and-pasted code into private functions.  Which was good, because they revealed a couple bugs:  sections of the redundant code had drifted in content from one another, so bits that were fixed in one turned out to not be fixed in the other.  Overall, saved 300 lines of code (!)
(The bug fixed here was that the line:

    fud->setUnitReferenceId(sr->getSpecies());

was not called for products (only reactants), so the upshot was that product units didn't get IDs that could be referenced.)

Also, since the original bug that started this hunt was due to re-used index variables for 'for' loops, made all the for loops in that function declare their index variables locally.  And while I was at it, I removed all uninitialized variable declarations in favor of local variables so that it's clear that they're not needed outside their local scope (and aren't left hanging around initialized to the last item on the list).