The settings for hyperref caused the PDF to always open on the front page, so when people quit their PDF viewer and restarted it, the viewer wouldn't go back to the page where they left.  Fixed.
This took a bit of experimentation, but I found the source of the problem.  It's an unexpected consequence of using an option that displays the bookmarks expanded when you first open the document.  The option is for the hyperref latex package and is set in sbmlpkgspec.cls.  The documentation for the option does not imply it would affect previewer's ability to reopen to the last page viewed, and at first I didn't think it did, but then testing in something other than Acrobat revealed that, sure enough, there's a difference between using the option and not using it.