Fix for flattening:  it is now possible (as it says in the spec) to explicitly delete something that otherwise was deleted by implication, i.e. delete both a species reference and the reaction it came from.  You can also replace a reaction and replace the species reference it contained.  Unfortunately, you cannot delete a reaction and replace a species reference it contained.  However, that may not be desirable anyway.
One issue with this change is that it changes the signature of some public functions that are used in flattening.  I saw no way to fix the above bug without changing the signature, since I needed to *not* delete things in those functions some times, based on information that function did not have.

I could theoretically give the new function arguments defaults, but would rather not--they are actually important for the new functionality.

Also, I discovered that Submodel::performDeletions wasn't ever used, and therefore was not tested, so I removed it.  This of course it another change to the API, but it seemed better to me to take out something that could go wrong than to leave it in.

Also also, Frank noticed that old versions of the STL had bugs where you couldn't declare defaults, so those functions have been split into the 'normal calls' and the 'protected calls with more arguments', which probably makes better functional sense anyway.

Finally, added a new validation rule to disallow replacing things that are children of deleted elements.  If this becomes legal, it means a headachey rewrite of the flattening code (again), but I think people on the list will likely agree that this situation is not helpful.