A fix, of sorts, for Frank's VS2013 bug:  if the ListOf::ListOf constructor creates an XMLNode out of itself, it will sometimes get it wrong, but the *next* time it needs to create an XMLNode out of itself, it will get it right!
In order for this to work, various qual ListOf*'s had to be changed so that if they were created with no namespaces, calling 'writeXMLNS' wouldn't crash. I put in the fix for some comp lists, too, but then stopped, as I wasn't sure whether the fix actually needed to go in, or if the qual lists were faulty to begin with for some reason--maybe the XMLNamespaces object was never supposed to be NULL?  Anyway.

Also, a few more #includes for strict includes to work.

If we actually want to use this fix, I recommend wrapping the ListOf constructor change in #ifdefs for VS2013.