Rewrote the top-level file reading loop so that it pays attention to #if/#ifdef/#ifndef/#else/#endif constructs.  It's far from a full preprocessor, but it handles the simple uses we tend to have in our files.  The symbols that it detects are defined using the command-line flag "--define".  By default, "SWIG" and "__cplusplus" are always defined.
This fixes a long-standing limitation in how swigdoc processed files.  It makes it possible to cleanly handle the math-legacy/math alternatives that are conditionalized on LIBSBML_USE_LEGACY_MATH.  More importantly, it also fixes a deeper problem that existed until now: if any source code in a .h or .i file was conditionalized with #if/#ifdef/#else, then the previous version of swigdoc could get the wrong version of a function/method/class definition because it read everything, without respecting the conditional at all.