Changed the core parser to use the new parsePackageInfix function; the old brackets/braces functions are now expunged entirely.
Also, changed vector<double*> to vector<double> because it's the same cost, and why?

Also also, added 'std::' to the vectors in the function in the .h files.  Turns out there were a couple spurious 'using namespace std' in the new AST*.h files; that was fixed in the trunk, but would reveal the problem here.